+++ app/src/main/java/com/example/weatherproj/ui/cities/CitiesScreen.kt
@@ -18,6 +18,9 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
 import com.example.weatherproj.FakeWeather
 import com.example.weatherproj.FakeWeatherData
 import com.example.weatherproj.R
